Seperate sass code
