+++ doc/colorizer.txt
@@ -351,14 +351,6 @@ Functions: ~
     |color_name_parser| - Grab all the colour values from
     `vim.api.nvim_get_color_map` and create a lookup table.
 
-    |sass_cleanup| - Cleanup sass variables
-
-    |sass_update_variables| - Parse the given lines for sass variabled and add
-    to SASS[buf].DEFINITIONS_ALL.
-
-    |sass_name_parser| - Parse the given line for sass color names
- check for value in SASS[buf].DEFINITIONS_ALL
-
     |rgb_function_parser| - Parse for rgb() css function and return rgb hex.
 
     |rgba_function_parser| - Parse for rgba() css function and return rgb hex.
@@ -398,53 +390,6 @@ color_name_parser({line}, {i}, {opts}) *colorizer.color_utils.color_name_parser*
 	{line} -  string: Line to parse
 	{i} -  number: Index of line from where to start parsing
 	{opts} -  table: Currently contains whether tailwind is enabled or not
-
-
-
-sass_cleanup({buf})                         *colorizer.color_utils.sass_cleanup*
-    Cleanup sass variables
-
-    Parameters: ~
-	{buf} -  number
-
-
-
-
-				   *colorizer.color_utils.sass_update_variables*
-sass_update_variables({buf}, {line_start}, {line_end}, {lines}, {color_parser},
-{options}, {options_local})
-    Parse the given lines for sass variabled and add to
-    SASS[buf].DEFINITIONS_ALL.
-
-     which is then used in |sass_name_parser|
-     If lines are not given, then fetch the lines with line_start and line_end
-
-
-    Parameters: ~
-	{buf} -  number
-	{line_start} -  number
-	{line_end} -  number
-	{lines} -  table|nil
-	{color_parser} -  function|boolean
-	{options} -  table: Buffer options
-	{options_local} -  table|nil: Buffer local variables
-
-    returns:~
-	boolean
-
-
-
-sass_name_parser({line}, {i}, {buf})    *colorizer.color_utils.sass_name_parser*
-    Parse the given line for sass color names
-     check for value in SASS[buf].DEFINITIONS_ALL
-
-    Parameters: ~
-	{line} -  string: Line to parse
-	{i} -  number: Index of line from where to start parsing
-	{buf} -  number
-
-    returns:~
-	number or nil, string or nil
 
 
 
@@ -565,6 +510,69 @@ make_matcher({options})                   *colorizer.matcher_utils.make_matcher*
     returns:~
 	function or boolean: function which will just parse the line for enabled
 	parsers
+
+
+
+==============================================================================
+SASS                                               *colorizer.sass-introduction*
+
+Helper functions to parse sass color variables
+
+
+==============================================================================
+LUA API                                                 *colorizer.sass-lua-api*
+
+Functions: ~
+    |sass_cleanup| - Cleanup sass variables and watch handlers
+
+    |sass_name_parser| - Parse the given line for sass color names
+ check for value in SASS[buf].DEFINITIONS_ALL
+
+    |sass_update_variables| - Parse the given lines for sass variabled and add
+    to SASS[buf].DEFINITIONS_ALL.
+
+
+sass_cleanup({buf})                                *colorizer.sass.sass_cleanup*
+    Cleanup sass variables and watch handlers
+
+    Parameters: ~
+	{buf} -  number
+
+
+
+sass_name_parser({line}, {i}, {buf})           *colorizer.sass.sass_name_parser*
+    Parse the given line for sass color names
+     check for value in SASS[buf].DEFINITIONS_ALL
+
+    Parameters: ~
+	{line} -  string: Line to parse
+	{i} -  number: Index of line from where to start parsing
+	{buf} -  number
+
+    returns:~
+	number or nil, string or nil
+
+
+
+
+					  *colorizer.sass.sass_update_variables*
+sass_update_variables({buf}, {line_start}, {line_end}, {lines}, {color_parser},
+{options}, {options_local})
+    Parse the given lines for sass variabled and add to
+    SASS[buf].DEFINITIONS_ALL.
+
+     which is then used in |sass_name_parser|
+     If lines are not given, then fetch the lines with line_start and line_end
+
+
+    Parameters: ~
+	{buf} -  number
+	{line_start} -  number
+	{line_end} -  number
+	{lines} -  table|nil
+	{color_parser} -  function|boolean
+	{options} -  table: Buffer options
+	{options_local} -  table|nil: Buffer local variables
