+++ doc/index.html
@@ -35,6 +35,7 @@
   <li><a href="modules/colorizer.buffer_utils.html">buffer_utils</a></li>
   <li><a href="modules/colorizer.color_utils.html">color_utils</a></li>
   <li><a href="modules/colorizer.matcher_utils.html">matcher_utils</a></li>
+  <li><a href="modules/colorizer.sass.html">sass</a></li>
   <li><a href="modules/colorizer.trie.html">trie</a></li>
   <li><a href="modules/utils.html">utils</a></li>
 </ul>
@@ -62,6 +63,10 @@
 	<tr>
 		<td class="name"  nowrap><a href="modules/colorizer.matcher_utils.html">colorizer.matcher_utils</a></td>
 		<td class="summary">Helper functions for colorizer to enable required parsers</td>
+	</tr>
+	<tr>
+		<td class="name"  nowrap><a href="modules/colorizer.sass.html">colorizer.sass</a></td>
+		<td class="summary">Helper functions to parse sass color variables</td>
 	</tr>
 	<tr>
 		<td class="name"  nowrap><a href="modules/colorizer.trie.html">colorizer.trie</a></td>
