+++ doc/modules/colorizer.buffer_utils.html
@@ -44,6 +44,7 @@
   <li><strong>buffer_utils</strong></li>
   <li><a href="../modules/colorizer.color_utils.html">color_utils</a></li>
   <li><a href="../modules/colorizer.matcher_utils.html">matcher_utils</a></li>
+  <li><a href="../modules/colorizer.sass.html">sass</a></li>
   <li><a href="../modules/colorizer.trie.html">trie</a></li>
   <li><a href="../modules/utils.html">utils</a></li>
 </ul>
