+++ doc/modules/colorizer.matcher_utils.html
@@ -42,6 +42,7 @@
   <li><a href="../modules/colorizer.buffer_utils.html">buffer_utils</a></li>
   <li><a href="../modules/colorizer.color_utils.html">color_utils</a></li>
   <li><strong>matcher_utils</strong></li>
+  <li><a href="../modules/colorizer.sass.html">sass</a></li>
   <li><a href="../modules/colorizer.trie.html">trie</a></li>
   <li><a href="../modules/utils.html">utils</a></li>
 </ul>
