+++ doc/modules/colorizer.sass.html
@@ -1,0 +1,188 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
+   "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+<head>
+    <title>colorizer Docs</title>
+    <link rel="stylesheet" href="../ldoc.css" type="text/css" />
+</head>
+<body>
+
+<div id="container">
+
+<div id="product">
+	<div id="product_logo"></div>
+	<div id="product_name"><big><b></b></big></div>
+	<div id="product_description"></div>
+</div> <!-- id="product" -->
+
+
+<div id="main">
+
+
+<!-- Menu -->
+
+<div id="navigation">
+<br/>
+<h1>colorizer</h1>
+
+<ul>
+  <li><a href="../index.html">Index</a></li>
+</ul>
+
+<h2>Contents</h2>
+<ul>
+<li><a href="#Functions">Functions</a></li>
+</ul>
+
+
+<h2>Modules</h2>
+<ul class="nowrap">
+  <li><a href="../modules/colorizer.html">colorizer</a></li>
+  <li><a href="../modules/colorizer.buffer_utils.html">buffer_utils</a></li>
+  <li><a href="../modules/colorizer.color_utils.html">color_utils</a></li>
+  <li><a href="../modules/colorizer.matcher_utils.html">matcher_utils</a></li>
+  <li><strong>sass</strong></li>
+  <li><a href="../modules/colorizer.trie.html">trie</a></li>
+  <li><a href="../modules/utils.html">utils</a></li>
+</ul>
+
+</div>
+
+<div id="content">
+
+<h1>Module <code>colorizer.sass</code></h1>
+<p>Helper functions to parse sass color variables</p>
+<p>
+
+</p>
+
+
+<h2><a href="#Functions">Functions</a></h2>
+<table class="function_list">
+	<tr>
+	<td class="name" nowrap><a href="#sass_cleanup">sass_cleanup (buf)</a></td>
+	<td class="summary">Cleanup sass variables and watch handlers</td>
+	</tr>
+	<tr>
+	<td class="name" nowrap><a href="#sass_name_parser">sass_name_parser (line, i, buf)</a></td>
+	<td class="summary">Parse the given line for sass color names
+ check for value in SASS[buf].DEFINITIONS_ALL</td>
+	</tr>
+	<tr>
+	<td class="name" nowrap><a href="#sass_update_variables">sass_update_variables (buf, line_start, line_end, lines, color_parser, options, options_local)</a></td>
+	<td class="summary">Parse the given lines for sass variabled and add to SASS[buf].DEFINITIONS_ALL.</td>
+	</tr>
+</table>
+
+<br/>
+<br/>
+
+
+    <h2 class="section-header "><a name="Functions"></a>Functions</h2>
+
+    <dl class="function">
+    <dt>
+    <a name = "sass_cleanup"></a>
+    <strong>sass_cleanup (buf)</strong>
+    </dt>
+    <dd>
+    Cleanup sass variables and watch handlers
+
+
+    <h3>Parameters:</h3>
+    <ul>
+        <li><span class="parameter">buf</span>
+         number
+        </li>
+    </ul>
+
+
+
+
+
+</dd>
+    <dt>
+    <a name = "sass_name_parser"></a>
+    <strong>sass_name_parser (line, i, buf)</strong>
+    </dt>
+    <dd>
+    Parse the given line for sass color names
+ check for value in SASS[buf].DEFINITIONS_ALL
+
+
+    <h3>Parameters:</h3>
+    <ul>
+        <li><span class="parameter">line</span>
+         string: Line to parse
+        </li>
+        <li><span class="parameter">i</span>
+         number: Index of line from where to start parsing
+        </li>
+        <li><span class="parameter">buf</span>
+         number
+        </li>
+    </ul>
+
+    <h3>Returns:</h3>
+    <ol>
+
+        number|nil, string|nil
+    </ol>
+
+
+
+
+</dd>
+    <dt>
+    <a name = "sass_update_variables"></a>
+    <strong>sass_update_variables (buf, line_start, line_end, lines, color_parser, options, options_local)</strong>
+    </dt>
+    <dd>
+    Parse the given lines for sass variabled and add to SASS[buf].DEFINITIONS<em>ALL.
+ which is then used in |sass</em>name<em>parser|
+ If lines are not given, then fetch the lines with line</em>start and line_end
+
+
+    <h3>Parameters:</h3>
+    <ul>
+        <li><span class="parameter">buf</span>
+         number
+        </li>
+        <li><span class="parameter">line_start</span>
+         number
+        </li>
+        <li><span class="parameter">line_end</span>
+         number
+        </li>
+        <li><span class="parameter">lines</span>
+         table|nil
+        </li>
+        <li><span class="parameter">color_parser</span>
+         function|boolean
+        </li>
+        <li><span class="parameter">options</span>
+         table: Buffer options
+        </li>
+        <li><span class="parameter">options_local</span>
+         table|nil: Buffer local variables
+        </li>
+    </ul>
+
+
+
+
+
+</dd>
+</dl>
+
+
+</div> <!-- id="content" -->
+</div> <!-- id="main" -->
+<div id="about">
+<i>generated by <a href="http://github.com/stevedonovan/LDoc">LDoc 1.4.6</a></i>
+<i style="float:right;">Last updated - September </i>
+</div> <!-- id="about" -->
+</div> <!-- id="container" -->
+</body>
+</html>
