+++ doc/modules/colorizer.trie.html
@@ -38,6 +38,7 @@
   <li><a href="../modules/colorizer.buffer_utils.html">buffer_utils</a></li>
   <li><a href="../modules/colorizer.color_utils.html">color_utils</a></li>
   <li><a href="../modules/colorizer.matcher_utils.html">matcher_utils</a></li>
+  <li><a href="../modules/colorizer.sass.html">sass</a></li>
   <li><strong>trie</strong></li>
   <li><a href="../modules/utils.html">utils</a></li>
 </ul>
